Extract URI parsing
