+++ phpstan-baseline.neon
@@ -130,7 +130,7 @@ parameters:
 			message: '#^Parameter \#1 \$callback of function array_map expects \(callable\(int\<0, 65535\>\|string\)\: mixed\)\|null, ''urldecode'' given\.$#'
 			identifier: argument.type
 			count: 1
-			path: src/Uri.php
+			path: src/UriParser.php
 
 		-
 			message: '#^Strict comparison using \=\=\= between non\-falsy\-string and '''' will always evaluate to false\.$#'
