+++ src/Uri.php
@@ -65,9 +65,9 @@ class Uri implements UriInterface, \JsonSerializable
     public function __construct(string $uri = '')
     {
         if ($uri !== '') {
-            $parts = self::parse($uri);
+            $parts = UriParser::parse($uri);
             if ($parts === false) {
-                throw new MalformedUriException("Unable to parse URI: $uri");
+                throw new MalformedUriException('Unable to parse URI: '.\addcslashes($uri, "\0..\37\177"));
             }
             try {
                 $this->applyParts($parts);
@@ -77,136 +77,6 @@ class Uri implements UriInterface, \JsonSerializable
                 throw new MalformedUriException($e->getMessage(), 0, $e);
             }
         }
-    }
-
-    /**
-     * UTF-8 aware \parse_url() replacement.
-     *
-     * The internal function produces broken output for non ASCII domain names
-     * (IDN) when used with locales other than "C".
-     *
-     * On the other hand, cURL understands IDN correctly only when UTF-8 locale
-     * is configured ("C.UTF-8", "en_US.UTF-8", etc.).
-     *
-     * @see https://bugs.php.net/bug.php?id=52923
-     * @see https://www.php.net/manual/en/function.parse-url.php#114817
-     * @see https://curl.se/libcurl/c/CURLOPT_URL.html#ENCODING
-     *
-     * @return array|false
-     */
-    private static function parse(string $url)
-    {
-        if (self::isPathNoSchemeReference($url)) {
-            return self::parsePathNoSchemeReference($url);
-        }
-
-        // Preserve bracketed IP-literals (IPv6 or IPvFuture) in scheme, userinfo,
-        // and network-path authorities before encoding. Userinfo is encoded
-        // separately so raw bytes cannot reach parse_url(), which mutates
-        // control characters instead of failing.
-        $prefix = '';
-        $ipv6Prefix = preg_match('%\A((?:[0-9A-Za-z+.-]+:)?//)(?:([^/?#@]*)(@))?(\[[^\]\x00-\x20\x7F/?#@]+\])(.*)\z%s', $url, $matches);
-
-        if ($ipv6Prefix === false) {
-            return false;
-        }
-
-        if ($ipv6Prefix === 1) {
-            /** @var array{0:string, 1:string, 2:string, 3:string, 4:string, 5:string} $matches */
-            $suffix = $matches[5];
-
-            // After the bracketed host only an optional numeric port and/or a
-            // path, query, or fragment may follow. Anything else (for example
-            // `:80@evil` or `:80x`) would let parse_url() reinterpret a
-            // different host.
-            if (preg_match('%\A(?::[0-9]*)?(?:[/?#].*)?\z%s', $suffix) !== 1) {
-                return false;
-            }
-
-            // RFC 3986 IP-literals contain no percent-encoding, so reject any
-            // "%" in the bracketed host rather than letting the urldecode()
-            // below turn an encoded octet into a different literal. This keeps
-            // parsing aligned with withHost()/Rfc3986::isValidHost().
-            if (str_contains($matches[4], '%')) {
-                return false;
-            }
-
-            $prefix = $matches[1];
-
-            if ($matches[3] === '@') {
-                /** @var string|null */
-                $encodedUserInfo = preg_replace_callback(
-                    '%[^:/@?&=#]+%usD',
-                    static function (array $matches): string {
-                        return urlencode($matches[0]);
-                    },
-                    $matches[2]
-                );
-
-                if ($encodedUserInfo === null) {
-                    return false;
-                }
-
-                $prefix .= $encodedUserInfo.'@';
-            }
-
-            $prefix .= $matches[4];
-            $url = $suffix;
-        }
-
-        /** @var string|null */
-        $encodedUrl = preg_replace_callback(
-            '%[^:/@?&=#]+%usD',
-            static function (array $matches): string {
-                return urlencode($matches[0]);
-            },
-            $url
-        );
-
-        if ($encodedUrl === null) {
-            return false;
-        }
-
-        $result = parse_url($prefix.$encodedUrl);
-
-        if ($result === false) {
-            return false;
-        }
-
-        return array_map('urldecode', $result);
-    }
-
-    private static function isPathNoSchemeReference(string $url): bool
-    {
-        if ($url === '' || str_starts_with($url, '/') || str_starts_with($url, '?') || str_starts_with($url, '#')) {
-            return false;
-        }
-
-        $firstSegment = substr($url, 0, strcspn($url, '/?#'));
-
-        return !str_contains($firstSegment, ':');
-    }
-
-    /**
-     * @return array{path: string, query?: string, fragment?: string}
-     */
-    private static function parsePathNoSchemeReference(string $url): array
-    {
-        $parts = [];
-
-        if (false !== ($fragmentPosition = strpos($url, '#'))) {
-            $parts['fragment'] = substr($url, $fragmentPosition + 1);
-            $url = substr($url, 0, $fragmentPosition);
-        }
-
-        if (false !== ($queryPosition = strpos($url, '?'))) {
-            $parts['query'] = substr($url, $queryPosition + 1);
-            $url = substr($url, 0, $queryPosition);
-        }
-
-        $parts['path'] = $url;
-
-        return $parts;
     }
 
     public function __toString(): string
@@ -478,7 +348,7 @@ class Uri implements UriInterface, \JsonSerializable
     public static function assertValidHost(string $host): void
     {
         if (!Rfc3986::isValidHost($host)) {
-            throw new \InvalidArgumentException(sprintf('Invalid host: "%s"', $host));
+            throw new \InvalidArgumentException(sprintf('Invalid host: "%s"', \addcslashes($host, "\0..\37\177")));
         }
     }
 
@@ -730,7 +600,7 @@ class Uri implements UriInterface, \JsonSerializable
         $scheme = Utils::asciiToLower($scheme);
 
         if (!Rfc3986::isValidScheme($scheme)) {
-            throw new \InvalidArgumentException(sprintf('Invalid scheme: "%s"', $scheme));
+            throw new \InvalidArgumentException(sprintf('Invalid scheme: "%s"', \addcslashes($scheme, "\0..\37\177")));
         }
 
         return $scheme;
@@ -830,7 +700,7 @@ class Uri implements UriInterface, \JsonSerializable
     private static function describeInvalidPort($port): string
     {
         if (\is_string($port)) {
-            return $port;
+            return \addcslashes($port, "\0..\37\177");
         }
 
         if (\is_int($port)) {
